refactor(authorization): Pull authorization out of common, into its own module
