+++ src/authorization/adapters/mod.rs
@@ -1,0 +1,2 @@
+pub mod basic_authorization;
+pub mod no_op;
