+++ src/authorization/domain/mod.rs
@@ -1,0 +1,3 @@
+pub mod action;
+pub mod error;
+pub mod resource;
