+++ src/authorization/mod.rs
@@ -1,0 +1,3 @@
+pub mod adapters;
+pub mod domain;
+pub mod ports;
