--- src/common/adapters/authorization_engine/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod basic_authorization;
-pub(crate) mod no_op;
