+++ src/common/adapters/mod.rs
@@ -1,2 +1,1 @@
-pub mod authorization_engine;
 pub mod project_repository;
