+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub(crate) mod authentication;
+pub mod authorization;
 pub(crate) mod common;
 pub(crate) mod potree_assets;
 pub mod potree_auth;
